+++ src/Justice.py
@@ -1,4 +1,6 @@
+from random import randint
 from jauge import *
+from legalite import *
 
 #Justice hérite de Jauge
 
@@ -39,3 +41,22 @@ class Justice(Jauge):
             self.justice = 0
         else :
             return self.justice
+
+    def chance_de_detection(self):
+        """Fonction qui gère la propabilité d'être detécté par la justice
+        NB : si la jauge justice est à 50%, cela ne veut pas dire que vous avez 1 chance / 2 d'être detécté(e), c'est un indiquateur qui plus il est élevé a de chance de vous envoyer en procès et d'aller en prison"""
+        just = Justice.get_jus()
+        if just >= 50 :                             #si la jauge de détéction par la justice passe au dessus de 50% :
+            if randint((just*0.1),10)== 10 :        #le joueur à selon son pourcentage entre 1 chance sur 10 d'être détécté ou 1 chance sur 2.
+                return Gestion.fin_jeu(2)           #fin du jeu
+        if just < 50 :
+            Justice.add_jus(-1)
+
+    def lien_justice_legalite(self):
+        """Cette fonction fait le lien entre la jauge de justice et de legalité
+        plus on est dans l'illégalité, plus on a de chance d'être detécté par la justice et aller en prison"""
+        legal = Legalite.get_leg()
+        if legal < 50 :             
+            Justice.add_jus(randint(2, legal-48)//4)    #Si le joueur a un pourcentage de légalité inférieur à 50, à chaques tour la jauge "justice" augmente petit à petit
+        if legal > 50 :                             #Si le joueur a un pourcentage de légalité supérieur à 50, son pourcentage de chance d'être détécté par la justice est mis à 10
+            Justice.set_jus(10)
